Fix empty "Flushing Volumes" dialogue
wxWidgets 3.1.5 is the version of wxWidgets that BambuStudio lists in
its builtin dependencies, and is the latest version of wxWidgets that
doesn't cause the empty "Flushing Volumes" dialogue that users saw with
version 3.2.4.

Downgrade to that version to fix the "Flushing Volumes" issue and
benefit from the app's own tests. Note that we had to backport a
number of fixes to make the build possible, including:
- upgrade to libwebkit2gtk-4.1 and libsoup-3.0
- fix -ldep_name=NOT-FOUND error in wx-config

Closes: #1